完善角色列表/添加/修改/删除,Validator分组验证优化，分页排序优化

+++ src/main/java/io/geekidea/springbootplus/system/service/impl/SysRoleServiceImpl.java
@@ -16,23 +16,24 @@
 
 package io.geekidea.springbootplus.system.service.impl;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.core.metadata.OrderItem;
+import com.baomidou.mybatisplus.extension.conditions.query.LambdaQueryChainWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.geekidea.springbootplus.framework.common.exception.BusinessException;
 import io.geekidea.springbootplus.framework.common.exception.DaoException;
 import io.geekidea.springbootplus.framework.common.exception.SpringBootPlusException;
+import io.geekidea.springbootplus.framework.common.service.impl.BaseServiceImpl;
 import io.geekidea.springbootplus.framework.pagination.PageUtil;
 import io.geekidea.springbootplus.framework.pagination.Paging;
-import io.geekidea.springbootplus.framework.common.service.impl.BaseServiceImpl;
-import io.geekidea.springbootplus.system.enums.StateEnum;
 import io.geekidea.springbootplus.system.convert.SysRoleConvert;
 import io.geekidea.springbootplus.system.entity.SysRole;
+import io.geekidea.springbootplus.system.enums.StateEnum;
 import io.geekidea.springbootplus.system.mapper.SysRoleMapper;
 import io.geekidea.springbootplus.system.param.sysrole.SysRolePageParam;
-import io.geekidea.springbootplus.system.param.sysrole.AddSysRoleParam;
-import io.geekidea.springbootplus.system.param.sysrole.UpdateSysRoleParam;
 import io.geekidea.springbootplus.system.param.sysrole.UpdateSysRolePermissionParam;
 import io.geekidea.springbootplus.system.service.SysPermissionService;
 import io.geekidea.springbootplus.system.service.SysRolePermissionService;
@@ -42,6 +43,7 @@ import io.geekidea.springbootplus.system.vo.SysRoleQueryVo;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.SetUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -79,49 +81,30 @@ public class SysRoleServiceImpl extends BaseServiceImpl<SysRoleMapper, SysRole> 
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public boolean saveSysRole(AddSysRoleParam addSysRoleParam) throws Exception {
-        String code = addSysRoleParam.getCode();
-        List<Long> permissionIds = addSysRoleParam.getPermissionIds();
+    public boolean saveSysRole(SysRole sysRole) throws Exception {
+        String code = sysRole.getCode();
         // 校验角色标识code唯一性
         if (isExistsByCode(code)) {
             throw new BusinessException("角色编码已存在");
         }
-        // 校验权限列表是否存在
-        if (!sysPermissionService.isExistsByPermissionIds(permissionIds)) {
-            throw new BusinessException("权限id不存在");
-        }
-
         // 保存角色
-        SysRole sysRole = SysRoleConvert.INSTANCE.addSysRoleParamToSysRole(addSysRoleParam);
         boolean saveRoleResult = super.save(sysRole);
         if (!saveRoleResult) {
             throw new DaoException("保存角色失败");
-        }
-
-        // 保存角色权限
-        Long roleId = sysRole.getId();
-        boolean saveRolePermissionResult = sysRolePermissionService.saveSysRolePermission(roleId, permissionIds);
-        if (!saveRolePermissionResult) {
-            throw new DaoException("保存角色权限失败");
         }
         return true;
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public boolean updateSysRole(UpdateSysRoleParam updateSysRoleParam) throws Exception {
-        Long roleId = updateSysRoleParam.getId();
+    public boolean updateSysRole(SysRole sysRole) throws Exception {
+        Long roleId = sysRole.getId();
         // 校验角色是否存在
-        SysRole sysRole = getById(roleId);
-        if (sysRole == null) {
+        if (getById(roleId) == null){
             throw new BusinessException("该角色不存在");
         }
         // 修改角色
-        sysRole.setName(updateSysRoleParam.getName())
-                .setType(updateSysRoleParam.getType())
-                .setRemark(updateSysRoleParam.getRemark())
-                .setState(updateSysRoleParam.getState())
-                .setUpdateTime(new Date());
+        sysRole.setUpdateTime(new Date());
         boolean updateResult = updateById(sysRole);
         if (!updateResult) {
             throw new DaoException("修改系统角色失败");
@@ -142,10 +125,15 @@ public class SysRoleServiceImpl extends BaseServiceImpl<SysRoleMapper, SysRole> 
         if (!deleteRoleResult) {
             throw new DaoException("删除角色失败");
         }
-        // 角色权限关系真实删除
-        boolean deletePermissionResult = sysRolePermissionService.deleteSysRolePermissionByRoleId(id);
-        if (!deletePermissionResult) {
-            throw new DaoException("删除角色权限关系失败");
+
+        // 判断角色是否有权限，如果有，则删除
+        boolean hasPermission = sysRolePermissionService.hasPermission(id);
+        if (hasPermission) {
+            // 角色权限关系真实删除
+            boolean deletePermissionResult = sysRolePermissionService.deleteSysRolePermissionByRoleId(id);
+            if (!deletePermissionResult) {
+                throw new DaoException("删除角色权限关系失败");
+            }
         }
         return true;
     }
@@ -153,7 +141,7 @@ public class SysRoleServiceImpl extends BaseServiceImpl<SysRoleMapper, SysRole> 
     @Override
     public SysRoleQueryVo getSysRoleById(Serializable id) throws Exception {
         SysRoleQueryVo sysRoleQueryVo = sysRoleMapper.getSysRoleById(id);
-        if (sysRoleQueryVo == null){
+        if (sysRoleQueryVo == null) {
             throw new SpringBootPlusException("角色不存在");
         }
         List<Long> permissionIds = sysRolePermissionService.getPermissionIdsByRoleId((Long) id);
@@ -163,8 +151,29 @@ public class SysRoleServiceImpl extends BaseServiceImpl<SysRoleMapper, SysRole> 
 
     @Override
     public Paging<SysRoleQueryVo> getSysRolePageList(SysRolePageParam sysRolePageParam) throws Exception {
-        Page page = PageUtil.getPage(sysRolePageParam);
-        IPage<SysRoleQueryVo> iPage = sysRoleMapper.getSysRolePageList(page, sysRolePageParam);
+        Page page = PageUtil.getPage(sysRolePageParam, OrderItem.desc(getLambdaColumn(SysRole::getCreateTime)));
+        // 此处演示单表，使用mybatisplus自带方法进行分页
+        LambdaQueryWrapper<SysRole> lambdaQueryWrapper = new LambdaQueryWrapper();
+        String keyword = sysRolePageParam.getKeyword();
+        String name = sysRolePageParam.getName();
+        String code = sysRolePageParam.getCode();
+        Integer state = sysRolePageParam.getState();
+        if (StringUtils.isNotBlank(keyword)) {
+            lambdaQueryWrapper
+                    .like(SysRole::getName, keyword)
+                    .or()
+                    .like(SysRole::getCode, keyword);
+        }
+        if (StringUtils.isNotBlank(name)) {
+            lambdaQueryWrapper.like(SysRole::getName, name);
+        }
+        if (StringUtils.isNotBlank(code)) {
+            lambdaQueryWrapper.like(SysRole::getCode, code);
+        }
+        if (state != null) {
+            lambdaQueryWrapper.eq(SysRole::getState, state);
+        }
+        IPage<SysRole> iPage = sysRoleMapper.selectPage(page, lambdaQueryWrapper);
         return new Paging(iPage);
     }
 
@@ -192,7 +201,7 @@ public class SysRoleServiceImpl extends BaseServiceImpl<SysRoleMapper, SysRole> 
         if (sysRole == null) {
             throw new BusinessException("该角色不存在");
         }
-        if (CollectionUtils.isNotEmpty(permissionIds)){
+        if (CollectionUtils.isNotEmpty(permissionIds)) {
             // 校验权限列表是否存在
             if (!sysPermissionService.isExistsByPermissionIds(permissionIds)) {
                 throw new BusinessException("权限列表id匹配失败");
@@ -212,18 +221,18 @@ public class SysRoleServiceImpl extends BaseServiceImpl<SysRoleMapper, SysRole> 
         log.debug("deleteSet = " + deleteSet);
         log.debug("addSet = " + addSet);
 
-        if (CollectionUtils.isNotEmpty(deleteSet)){
+        if (CollectionUtils.isNotEmpty(deleteSet)) {
             // 删除权限关联
             UpdateWrapper updateWrapper = new UpdateWrapper();
-            updateWrapper.eq("role_id",roleId);
-            updateWrapper.in("permission_id",deleteSet);
+            updateWrapper.eq("role_id", roleId);
+            updateWrapper.in("permission_id", deleteSet);
             boolean deleteResult = sysRolePermissionService.remove(updateWrapper);
             if (!deleteResult) {
                 throw new DaoException("删除角色权限关系失败");
             }
         }
 
-        if (CollectionUtils.isNotEmpty(addSet)){
+        if (CollectionUtils.isNotEmpty(addSet)) {
             // 新增权限关联
             boolean addResult = sysRolePermissionService.saveSysRolePermissionBatch(roleId, addSet);
             if (!addResult) {
